Re-worked directionality
Once tagged is scanned, reader switches from antenna 1 to antenna 2 and added a timer for antenna 1 so that if the bird is at the box entrance it will switch to antenna 2 but then switch back to antenna 1. Still need to work out what happens if another bird shows up at entrance if there is already a bird in the box.